More work on categories
