+++ src/app/deals/[category]/page.tsx
@@ -1,6 +1,6 @@
 import { DealsRecord, getXataClient } from '@/xata';
-import Deals from '../../DealsList';
-import DealsFilter from '@/components/DealsFilter';
+import CategoryOptions from '@/components/CategoryOptions';
+import DealsList from '../../DealsList';
 
 export default async function CategoryPage({
   params,
@@ -15,7 +15,10 @@ export default async function CategoryPage({
     .getAll();
   return (
     <div>
-      <DealsFilter deals={deals} />
+      <h1 className="text-4xl text-center font-bold mb-10 text-white">
+        Top <span className="text-teal-500">{params.category}</span> Deals
+      </h1>
+      <DealsList deals={deals} />
     </div>
   );
 }
